Update azuredeploy.json
The spaces in all of the NSg rules --- lead to the below error 
Removed all the spaces in this PR 

The deployment 'azuredeploy' failed with error(s). Showing 1 out of 1 error(s). Status Message: Resource name Azure Portal is invalid. The name can be up to 80 characters long. It must begin with a word character, and it must end with a word character or with '_'. The name may contain word characters or '.', '-', '_'.